+++ edc-extension4aas/src/main/java/de/fraunhofer/iosb/app/Endpoint.java
@@ -464,7 +464,7 @@ public class Endpoint {
      */
     private void removeAssetsContracts(List<? extends IdsAssetElement> elements) {
         elements.forEach(element -> {
-            resourceController.deleteAsset(element.getIdsAssetId());
+            resourceController.deleteAssetAndContracts(element.getIdsAssetId());
             resourceController.deleteContract(element.getIdsContractId());
         });
     }